Handle cycle in suppressed exceptions
With the added support for suppressed exceptions in ExceptionPlaceholder
the code needs to handle when those have a cycle, which has a higher
chance of happening than a cycle in exception cause.
This commit makes sure we attempt to detect such a cycle and break the
exception chain during serialization if that is the case.

Fixes #9487